C4DtoA Support Added
Added Cinema4D resource files for the shader. Updated the readme file
with additional installation path options.
