+++ C4DtoA/res/description/aitag_vr_camera.res
@@ -1,0 +1,50 @@
+CONTAINER AITAG_VR_CAMERA
+{
+   NAME aitag_vr_camera;
+
+   GROUP C4DAI_VR_CAMERA_MAIN_GRP
+   {
+        DEFAULT 1;
+      
+        GROUP C4DAI_VR_CAMERA_MAIN_ATTRIBUTES_GRP
+        {
+            DEFAULT 1;
+
+            AIPARAM C4DAIP_VR_CAMERA_MODE {}
+            AIPARAM C4DAIP_VR_CAMERA_PROJECTION {}
+            AIPARAM C4DAIP_VR_CAMERA_EYESEPARATION {}
+            AIPARAM C4DAIP_VR_CAMERA_EYETONECKDISTANCE {}
+        }
+        
+        GROUP C4DAI_VR_CAMERA_POLE_MERGEING_GRP
+        {
+            DEFAULT 1;
+            
+            AIPARAM C4DAIP_VR_CAMERA_TOPMERGEMODE {}
+            AIPARAM C4DAIP_VR_CAMERA_TOPMERGEANGLE {}
+            AIPARAM C4DAIP_VR_CAMERA_TOPMERGEEXP {}
+            AIPARAM C4DAIP_VR_CAMERA_BOTTOMMERGEMODE {}
+            AIPARAM C4DAIP_VR_CAMERA_BOTTOMMERGEANGLE {}
+            AIPARAM C4DAIP_VR_CAMERA_BOTTOMMERGEEXP {}
+            AIPARAM C4DAIP_VR_CAMERA_MERGESHADER {}
+        }
+        
+        GROUP C4DAI_VR_CAMERA_EXTRAS_GRP
+        {
+            DEFAULT 1;
+
+            SEPARATOR {LINE;}
+            AIPARAM C4DAIP_VR_CAMERA_EXPOSURE {}
+            AIPARAM C4DAIP_VR_CAMERA_FILTERMAP {}
+            SEPARATOR {LINE;}
+            AIPARAM C4DAIP_VR_CAMERA_ROLLING_SHUTTER {}
+            AIPARAM C4DAIP_VR_CAMERA_ROLLING_SHUTTER_DURATION {}
+            SEPARATOR {LINE;}
+            AIPARAM C4DAIP_VR_CAMERA_SHUTTER_START {}
+            AIPARAM C4DAIP_VR_CAMERA_SHUTTER_END {}
+            AIPARAM C4DAIP_VR_CAMERA_SHUTTER_TYPE {}
+            AIPARAM C4DAIP_VR_CAMERA_SHUTTER_CURVE {}
+        }
+   }
+}
+
