--- VRCamera.mtd
@@ -1,62 +1,0 @@
-# THIRDPARTY_NODEID_VR_CAMERA      0x00116500
-[node vr_camera]
-    maya.name               STRING  "camera"
-    maya.id                 INT     0x00116500
-    maya.translator         STRING  "vr_camera"
-     
-    [attr mode]
-        maya.name           STRING  "aiMode"
-        default             INT     0
-     
-    [attr projection]
-        maya.name           STRING  "aiProjection"
-        default             INT     0
-
-    [attr eyeSeparation]
-        maya.name           STRING  "aiEyeSeparation"
-        softmin             FLOAT   0
-        softmax             FLOAT   1
-        default             FLOAT   0.65
-
-    [attr eyeToNeckDistance]
-        maya.name           STRING  "aiEyeToNeckDistance"
-        softmin             FLOAT   0
-        softmax             FLOAT   1
-        default             FLOAT   0.0
-
-    [attr topMergeMode]
-        maya.name           STRING  "aiTopMergeMode"
-        default             INT     2
-     
-    [attr topMergeAngle]
-        maya.name           STRING  "aiTopMergeAngle"
-        softmin             FLOAT   0
-        max                 FLOAT   90
-        default             FLOAT   0
-        
-    [attr topMergeExp]
-        maya.name           STRING  "aiTopMergeExp"
-        min                 FLOAT   0
-        softmax             FLOAT   4
-        default             FLOAT   1
-        
-    [attr bottomMergeMode]
-        maya.name           STRING  "aiBottomMergeMode"
-        default             INT     2
-     
-    [attr bottomMergeAngle]
-        maya.name           STRING  "aiBottomMergeAngle"
-        min                 FLOAT   -90
-        softmax             FLOAT   0
-        default             FLOAT   0
-        
-    [attr bottomMergeExp]
-        maya.name           STRING  "aiBottomMergeExp"
-        min                 FLOAT   0
-        softmax             FLOAT   4
-        default             FLOAT   1
-
-    [attr mergeShader]
-        maya.name           STRING  "aiMergeShader"
-        linkable            BOOL    True
-
